+++ Inventory Control/Dados/DeletarDadosFornecedor.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Guna.UI.WinForms;
+using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Data.SqlClient;
@@ -11,22 +12,40 @@ namespace Inventory_Control.Dados
 {
     internal class DeletarDadosFornecedor : ConectarBanco
     {
-        public void DeletarFornecedor(string _cNPJ)
+        private VerificacaoDeExistencia VC = new VerificacaoDeExistencia();
+        private BuscarDadosFornecedor BF = new BuscarDadosFornecedor();
+
+        public void DeletarFornecedor(string _cNPJ, GunaDataGridView _tabela)
         {
             try
             {
-                using (SqlConnection conexaoSQL = AbrirConexao())
+                if (VC.BuscarExistenciaFornecedor(_cNPJ))
                 {
-                    string query = "delete from Fornecedor where CNPJ = @CNPJ";
+                    using (SqlConnection conexaoSQL = AbrirConexao())
+                    {
+                        BF.BuscarFornecedor(_cNPJ, _tabela);
 
-                    SqlCommand cmd = new SqlCommand(query, conexaoSQL);
-                    cmd.Parameters.Add("@CNPJ", SqlDbType.VarChar).Value = _cNPJ;
-                    cmd.ExecuteNonQuery();
+                        DialogResult OpcaoDoUsuario = new DialogResult();
+                        OpcaoDoUsuario = MessageBox.Show("Deseja Excluir o Fornecedor?", "Atenção!", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+                        if (OpcaoDoUsuario == DialogResult.Yes)
+                        {
+                            string query = "delete from Fornecedor where CNPJ = @CNPJ";
+                            SqlCommand cmd = new SqlCommand(query, conexaoSQL);
+                            cmd.Parameters.AddWithValue("@cNPJ", _cNPJ);
+                            cmd.ExecuteNonQuery();
+
+                            MessageBox.Show("Fornecedor Excluido Com Sucesso", "Informação!", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        }
+                    }
+                }
+                else
+                {
+                    MessageBox.Show("Fornecedor Não Encontrado!", "Informação!", MessageBoxButtons.OK, MessageBoxIcon.Information);
                 }
             }
             catch (Exception x)
             {
-                MessageBox.Show(x.Message.ToString());
+                MessageBox.Show(x.ToString());
             }
         }
     }
